Add FileFormAdminMixin
